+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
 - 2026-??-?? 2.3.23
-    - tidied BLEDevice
+    - XS BLE
+        - tidied BLEDevice
+        - added USBCPowerConnectionState
+        - added SpectrometerState.power_connection_state
+        - added SpectrometerSettings.cpu_unique_id
+        - added SET_POWER_OFF, SET_RESET_UNIT, GET/SET_AUTO_RAMAN_PARAMS, 
+          GET_IMAGE_SENSOR_STATE, GET_CPU_UNIQUE_ID, GET_USB_ADAPTER_INFO, 
+          GET/SET_LASER_OFF_DELAY_MS
 - 2026-04-10 2.3.22
     - added Reading.protocol
     - XS